Fix 'Instable vertical position of alternate combined fragment widget'
The reason for this bug was that some of the variables used for
calculating the vertical position are of type int, which cuts the
fragment value of the vertical position.

BUG:390677
FIXED-IN:2.24.3 (KDE Applications 17.12.2)

+++ umbrello/umlwidgets/combinedfragmentwidget.cpp
@@ -71,7 +71,7 @@ void CombinedFragmentWidget::paint(QPainter *painter, const QStyleOptionGraphics
     int w = width();
     int h = height();
     int line_width = initialLabelWidth;
-    int old_Y;
+    qreal old_Y;
 
     setPenFromSettings(painter);
